- separate packages for server and event listener and extract this code from 'main' - utils package to store all utils (like hashing, validations, etc.) - some small cleaning + comments
