+++ decentralized-api/internal/event_listener/event_listener.go
@@ -1,4 +1,4 @@
-package main
+package event_listener
 
 import (
 	"context"
@@ -6,6 +6,7 @@ import (
 	"decentralized-api/broker"
 	"decentralized-api/chainevents"
 	cosmosclient "decentralized-api/cosmosclient"
+	"decentralized-api/internal/server"
 	"decentralized-api/poc"
 	"decentralized-api/upgrade"
 	"encoding/json"
@@ -81,7 +82,7 @@ func StartEventListener(
 				slog.Warn("Websocket connection closed", "errorType", fmt.Sprintf("%T", err), "error", err)
 				if upgrade.CheckForUpgrade(configManager) {
 					slog.Error("Upgrade required! Exiting...")
-					panic("Upgrade required")
+					panic("Upgrade required") // TODO: better not use panic in code
 				}
 				continue
 			}
@@ -91,8 +92,10 @@ func StartEventListener(
 		var event chainevents.JSONRPCResponse
 		if err = json.Unmarshal(message, &event); err != nil {
 			slog.Error("Error unmarshalling message to JSONRPCResponse", "error", err, "message", message)
+			continue // no sense to check event, if it wasn't unmarshalled correctly
 		}
 
+		// TODO: goroutines are created out of control, it can use crazy amount of recourses
 		go func() {
 			switch event.Result.Data.Type {
 			case "tendermint/event/NewBlock":
@@ -106,28 +109,6 @@ func StartEventListener(
 			}
 		}()
 	}
-}
-
-func getParams(ctx context.Context, transactionRecorder cosmosclient.InferenceCosmosClient) (*types.QueryParamsResponse, error) {
-	var params *types.QueryParamsResponse
-	var err error
-	for i := 0; i < 10; i++ {
-		params, err = transactionRecorder.NewInferenceQueryClient().Params(ctx, &types.QueryParamsRequest{})
-		if err == nil {
-			return params, nil
-		}
-
-		if strings.HasPrefix(err.Error(), "rpc error: code = Unknown desc = inference is not ready") {
-			slog.Info("Inference not ready, retrying...", "attempt", i+1, "error", err)
-			time.Sleep(2 * time.Second) // Try a longer wait for specific inference delays
-			continue
-		}
-		// If not an RPC error, log and return early
-		slog.Error("Failed to get chain params", "error", err)
-		return nil, err
-	}
-	slog.Error("Exhausted all retries to get chain params", "error", err)
-	return nil, err
 }
 
 func handleMessage(
@@ -150,9 +131,9 @@ func handleMessage(
 	//}
 	switch action {
 	case finishInferenceAction:
-		SampleInferenceToValidate(event.Result.Events["inference_finished.inference_id"], transactionRecorder, nodeBroker, currentConfig)
+		server.SampleInferenceToValidate(event.Result.Events["inference_finished.inference_id"], transactionRecorder, nodeBroker, currentConfig)
 	case validationAction:
-		VerifyInvalidation(event.Result.Events, transactionRecorder, nodeBroker)
+		server.VerifyInvalidation(event.Result.Events, transactionRecorder, nodeBroker)
 	}
 }
 
@@ -193,4 +174,26 @@ func getWebsocketUrl(config *apiconfig.Config) string {
 
 	// Construct the new URL
 	return u.String()
+}
+
+func GetParams(ctx context.Context, transactionRecorder cosmosclient.InferenceCosmosClient) (*types.QueryParamsResponse, error) {
+	var params *types.QueryParamsResponse
+	var err error
+	for i := 0; i < 10; i++ {
+		params, err = transactionRecorder.NewInferenceQueryClient().Params(ctx, &types.QueryParamsRequest{})
+		if err == nil {
+			return params, nil
+		}
+
+		if strings.HasPrefix(err.Error(), "rpc error: code = Unknown desc = inference is not ready") {
+			slog.Info("Inference not ready, retrying...", "attempt", i+1, "error", err)
+			time.Sleep(2 * time.Second) // Try a longer wait for specific inference delays
+			continue
+		}
+		// If not an RPC error, log and return early
+		slog.Error("Failed to get chain params", "error", err)
+		return nil, err
+	}
+	slog.Error("Exhausted all retries to get chain params", "error", err)
+	return nil, err
 }
