+++ decentralized-api/internal/event_listener/event_listener_test.go
@@ -1,4 +1,4 @@
-package main
+package event_listener
 
 import (
 	"decentralized-api/chainevents"
