+++ decentralized-api/internal/server/server_test.go
@@ -1,4 +1,4 @@
-package main
+package server
 
 import (
 	"decentralized-api/completionapi"
